--- src/interface_adapter/ReturnToPreviousView/ReturnPresenter.java
@@ -1,4 +1,0 @@
-package interface_adapter.ReturnToPreviousView;
-
-public class ReturnPresenter {
-}
