+++ src/interface_adapter/return_home/ViewChangeObserver.java
@@ -1,4 +1,4 @@
-package interface_adapter.ReturnToPreviousView;
+package interface_adapter.return_home;
 
 public interface ViewChangeObserver {
     void onViewChange(String viewName);
