--- src/use_case/ReturnToPreviousView/ReturnDataAccessInterface.java
@@ -1,4 +1,0 @@
-package use_case.ReturnToPreviousView;
-
-public interface ReturnDataAccessInterface {
-}
