--- src/use_case/ReturnToPreviousView/ReturnInputBoundary.java
@@ -1,7 +1,0 @@
-package use_case.ReturnToPreviousView;
-
-import use_case.signup.SignupInputData;
-
-public interface ReturnInputBoundary {
-    void navigateBack();
-}
