--- src/use_case/ReturnToPreviousView/ReturnInputData.java
@@ -1,4 +1,0 @@
-package use_case.ReturnToPreviousView;
-
-public class ReturnInputData {
-}
