--- src/use_case/ReturnToPreviousView/ReturnOutputData.java
@@ -1,4 +1,0 @@
-package use_case.ReturnToPreviousView;
-
-public class ReturnOutputData {
-}
