+++ src/use_case/return_home/ReturnInputBoundary.java
@@ -1,0 +1,5 @@
+package use_case.return_home;
+
+public interface ReturnInputBoundary {
+    void navigateBack();
+}
