+++ src/use_case/return_home/ReturnInputData.java
@@ -1,0 +1,4 @@
+package use_case.return_home;
+
+public class ReturnInputData {
+}
