+++ src/use_case/return_home/ReturnOutputBoundary.java
@@ -1,4 +1,4 @@
-package use_case.ReturnToPreviousView;
+package use_case.return_home;
 
 public interface ReturnOutputBoundary {
 }
