+++ src/use_case/return_home/ReturnOutputData.java
@@ -1,0 +1,4 @@
+package use_case.return_home;
+
+public class ReturnOutputData {
+}
